requirements.txt
# API Framework
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0

# Machine Learning and LLMs
torch>=2.0.1
transformers>=4.33.2
accelerate>=0.23.0
bitsandbytes>=0.41.1
optimum>=1.13.1
auto-gptq>=0.5.0
sentencepiece>=0.1.99
einops>=0.6.1
onnxruntime>=1.15.1
onnxruntime-gpu>=1.15.1; sys_platform != 'darwin' or platform_machine != 'arm64'

# UI
gradio>=3.41.2

# Utils
python-dotenv>=1.0.0
pyyaml>=6.0.1
loguru>=0.7.0
psutil>=5.9.5
huggingface_hub>=0.17.3
tqdm>=4.66.1
numpy>=1.24.3
pandas>=2.0.3

# Development and testing
pytest>=7.4.2
black>=23.9.1
isort>=5.12.0
pylint>=2.17.5